# openapi.yaml
components:
  schemas:
...
    CreateDataset:
      description: A Dataset is a collection of Files that is prepared for distribution
        and is tied to a Data Access Policy.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        description:
          description: Description of an entity.
          title: Description
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        has_analysis:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateAnalysis'
            type: array
          - items:
              type: string
            type: array
          description: One or more Analysis entities that are referenced by this Dataset.
          title: Has Analysis
        has_attribute:
          description: Key/value pairs corresponding to an entity.
          items:
            $ref: '#/components/schemas/metadata_repository_service__creation_models__Attribute'
          title: Has Attribute
          type: array
        has_data_access_policy:
          anyOf:
          - $ref: '#/components/schemas/CreateDataAccessPolicy'
          - type: string
          description: The Data Access Policy that applies to this Dataset.
          title: Has Data Access Policy
        has_experiment:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateExperiment'
            type: array
          - items:
              type: string
            type: array
          description: One or more Experiment entities that are referenced by this
            Dataset.
          title: Has Experiment
        has_file:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateFile'
            type: array
          - items:
              type: string
            type: array
          description: One or more File entities that collectively are part of this
            Dataset.
          title: Has File
        has_publication:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreatePublication'
            type: array
          - items:
              type: string
            type: array
          description: One or more Publication entities associated with this Dataset.
          title: Has Publication
        has_sample:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateSample'
            type: array
          - items:
              type: string
            type: array
          description: One or more Sample entities that are referenced by this Dataset.
          title: Has Sample
        has_study:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateStudy'
            type: array
          - items:
              type: string
            type: array
          description: One or more Study entities that are referenced by this Dataset.
          title: Has Study
        release_date:
          description: The timestamp (in ISO 8601 format) when the entity was released
            for public consumption.
          title: Release Date
          type: string
        release_status:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__creation_models__ReleaseStatusEnum'
          description: The release status of a Dataset.
        schema_type:
          enum:
          - CreateDataset
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        title:
          description: A title for the submitted Dataset.
          title: Title
          type: string
        type:
          description: The type of a dataset.
          items:
            type: string
          title: Type
          type: array
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: CreateDataset
      type: object
...
    Dataset:
      description: A Dataset is a collection of Files that is prepared for distribution
        and is tied to a Data Access Policy.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        description:
          description: Description of an entity.
          title: Description
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        has_analysis:
          anyOf:
          - items:
              $ref: '#/components/schemas/Analysis'
            type: array
          - items:
              type: string
            type: array
          description: One or more Analysis entities that are referenced by this Dataset.
          title: Has Analysis
        has_attribute:
          description: Key/value pairs corresponding to an entity.
          items:
            $ref: '#/components/schemas/metadata_repository_service__models__Attribute'
          title: Has Attribute
          type: array
        has_data_access_policy:
          anyOf:
          - $ref: '#/components/schemas/DataAccessPolicy'
          - type: string
          description: The Data Access Policy that applies to this Dataset.
          title: Has Data Access Policy
        has_experiment:
          anyOf:
          - items:
              $ref: '#/components/schemas/Experiment'
            type: array
          - items:
              type: string
            type: array
          description: One or more Experiment entities that are referenced by this
            Dataset.
          title: Has Experiment
        has_file:
          anyOf:
          - items:
              $ref: '#/components/schemas/File'
            type: array
          - items:
              type: string
            type: array
          description: One or more File entities that collectively are part of this
            Dataset.
          title: Has File
        has_publication:
          anyOf:
          - items:
              $ref: '#/components/schemas/Publication'
            type: array
          - items:
              type: string
            type: array
          description: One or more Publication entities associated with this Dataset.
          title: Has Publication
        has_sample:
          anyOf:
          - items:
              $ref: '#/components/schemas/Sample'
            type: array
          - items:
              type: string
            type: array
          description: One or more Sample entities that are referenced by this Dataset.
          title: Has Sample
        has_study:
          anyOf:
          - items:
              $ref: '#/components/schemas/Study'
            type: array
          - items:
              type: string
            type: array
          description: One or more Study entities that are referenced by this Dataset.
          title: Has Study
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        release_date:
          description: The timestamp (in ISO 8601 format) when the entity was released
            for public consumption.
          title: Release Date
          type: string
        release_status:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__ReleaseStatusEnum'
          description: The release status of a Dataset.
        schema_type:
          enum:
          - Dataset
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        title:
          description: A title for the submitted Dataset.
          title: Title
          type: string
        type:
          description: The type of a dataset.
          items:
            type: string
          title: Type
          type: array
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Dataset
      type: object
    DatasetStatusPatch:
      description: An object that can be used to change the release status of a Dataset.
      properties:
        release_status:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__ReleaseStatusEnum'
          description: The release status of a Dataset.
      title: DatasetStatusPatch
      type: object
    DatasetSummary:
      description: Dataset summary
      properties:
        accession:
          description: GHGA Accession of the Dataset.
          title: Accession
          type: string
        dac_email:
          description: DAC contact email
          title: Dac Email
          type: string
        description:
          description: Description of the Dataset.
          title: Description
          type: string
        ega_accession:
          description: EGA Accession of the Dataset.
          title: Ega Accession
          type: string
        experiment_summary:
          allOf:
          - $ref: '#/components/schemas/Summary'
          description: Experiment summary
          title: Experiment Summary
        file_summary:
          allOf:
          - $ref: '#/components/schemas/Summary'
          description: File summary
          title: File Summary
        id:
          description: Dataset ID
          title: Id
          type: string
        sample_summary:
          allOf:
          - $ref: '#/components/schemas/Summary'
          description: Sample summary
          title: Sample Summary
        study_summary:
          allOf:
          - $ref: '#/components/schemas/Summary'
          description: Study summary
          title: Study Summary
        title:
          description: A title for the submitted Dataset.
          title: Title
          type: string
        type:
          description: The type of the Dataset.
          items:
            type: string
          title: Type
          type: array
      title: DatasetSummary
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Summary:
      description: Summary
      properties:
        count:
          description: count of the summary item
          title: Count
          type: integer
        stats:
          description: a dictionary of stats for the summary item
          title: Stats
          type: object
      title: Summary
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    metadata_repository_service__creation_models__Attribute:
      description: A key/value pair that further characterizes an entity.
      properties:
        key:
          description: The key for an attribute.
          title: Key
          type: string
        key_type:
          description: A semantic type that characterizes the attribute key. Usually
            this is a term from an ontology. For example, 'MAXO:0000616' indicates
            that the attribute is a measurement of oxygen saturation in the blood.
          title: Key Type
          type: string
        value:
          description: The value for an attribute. Usually this is a numerical value
            (without the units).
          title: Value
          type: string
        value_type:
          description: The value type that characterizes the attribute value. Usually
            this is a term from an ontology that describes how to interpret the value.
            For example, 'SIO:001413' indicates that the value is to be interpreted
            as a percentage.
          title: Value Type
          type: string
      title: Attribute
      type: object
...
    metadata_repository_service__creation_models__ReleaseStatusEnum:
      description: An enumeration.
      enum:
      - unreleased
      - released
      title: ReleaseStatusEnum
      type: string
...
    metadata_repository_service__models__ReleaseStatusEnum:
      description: An enumeration.
      enum:
      - unreleased
      - released
      title: ReleaseStatusEnum
      type: string
...
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /dataset_summary/{dataset_id}:
    get:
      description: Given a Dataset ID, get the Dataset summary from the metadata store.
      operationId: get_dataset_summary_dataset_summary__dataset_id__get
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Dataset summary
      tags:
      - Query
  /datasets:
    post:
      description: 'Given a list of File accessions and a DataAccessPolicy accession,
        create a

        Dataset and write to the metadata store.'
      operationId: create_datasets_datasets_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDataset'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a Dataset
      tags:
      - Dataset
  /datasets/{dataset_accession}:
    patch:
      description: Update status of a Dataset entity.
      operationId: update_dataset_status_datasets__dataset_accession__patch
      parameters:
      - in: path
        name: dataset_accession
        required: true
        schema:
          title: Dataset Accession
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetStatusPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update status of a Dataset
      tags:
      - Dataset
  /datasets/{dataset_id}:
    get:
      description: Given a Dataset ID, get the Dataset record from the metadata store.
      operationId: get_datasets_datasets__dataset_id__get
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          title: Dataset Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Dataset
      tags:
      - Query
  /experiment_processes/{experiment_process_id}:
    get:
      description: Given a ExperimentProcess ID, get the ExperimentProcess record
        from the metadata store.
      operationId: get_experiment_processes_experiment_processes__experiment_process_id__get
      parameters:
      - in: path
        name: experiment_process_id
        required: true
        schema:
          title: Experiment Process Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentProcess'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a ExperimentProcess
      tags:
      - Query
  /experiments/{experiment_id}:
    get:
      description: Given a Experiment ID, get the Experiment record from the metadata
        store.
      operationId: get_experiments_experiments__experiment_id__get
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          title: Experiment Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get an Experiment
      tags:
      - Query
...